refactor

+++ src/UserDesc.h
@@ -8,8 +8,8 @@ namespace efs {
 struct UserDesc {
     std::string user;
     std::string password;
-    int32_t uid;
-    int32_t gid;
+    int16_t uid;
+    int16_t gid;
     std::string root_path;
 
     UserDesc();
